ep bench: split per-backend setup into separate modules; clarify review comments
- Move setup_mscclpp/nccl/deepep/flashinfer into ep_bench_{mscclpp,nccl,deepep,flashinfer}.py and shared helpers into ep_bench_common.py; run_ep_bench_python.py now imports them (addresses review: split backends into separate files for readability).
- Clarify --ep-layout native help (native = no cross-backend normalization; each backend keeps its own default layout).
- Document the NCCL-EP JIT source/include-dir knobs (why they exist: locate device sources/headers for the runtime JIT; defaults match the in-tree build).
- Expand the mscclpp cuda-graph comment explaining why dispatch/combine use two separate graphs (per-phase kineto dispatch-vs-combine timing) rather than one combined graph.
