+++ test/python/ep/run_ep_bench_python.py
@@ -113,7 +113,6 @@ from __future__ import annotations
 
 import argparse
 import ctypes
-import gc
 import glob
 import os
 import subprocess
@@ -121,6 +120,17 @@ import time
 
 import torch
 from mpi4py import MPI
+
+from ep_bench_common import (
+    init_mpi,
+    make_inputs,
+    _init_torch_nccl,
+    _mpi_stats,
+)
+from ep_bench_mscclpp import setup_mscclpp
+from ep_bench_nccl import setup_nccl
+from ep_bench_deepep import setup_deepep
+from ep_bench_flashinfer import setup_flashinfer
 
 
 # ----------------------------------------------------------------------------
@@ -167,11 +177,14 @@ def parse_args() -> argparse.Namespace:
         "--ep-layout",
         choices=["native", "rank_major", "expert_major"],
         default="native",
-        help="received-token dispatch layout. 'native' keeps each backend's current default "
-        "(nccl=expert_major, mscclpp=expert_major, deepep=rank_major, flashinfer=rank_major). "
-        "'rank_major'/'expert_major' override where supported: nccl (Layout.RANK_MAJOR/EXPERT_MAJOR) "
-        "and deepep (rank_major=plain, expert_major=do_expand). mscclpp LL is expert-major only and "
-        "flashinfer is rank-major only; an unsupported request is noted and the native layout is kept.",
+        help="received-token dispatch layout. 'native' means no cross-backend normalization: each "
+        "backend keeps the layout its own default code path produces "
+        "(nccl=expert_major, mscclpp=expert_major, deepep=rank_major, flashinfer=rank_major), which is "
+        "the apples-to-apples default for comparing each backend's out-of-the-box behavior. "
+        "'rank_major'/'expert_major' force a specific layout where supported: nccl "
+        "(Layout.RANK_MAJOR/EXPERT_MAJOR) and deepep (rank_major=plain, expert_major=do_expand). "
+        "mscclpp LL is expert-major only and flashinfer is rank-major only; an unsupported request is "
+        "noted and the native layout is kept.",
     )
     p.add_argument(
         "--validate",
@@ -186,7 +199,10 @@ def parse_args() -> argparse.Namespace:
         "kernel timer is torch kineto (EP_KERNEL_TIMER=kineto) with a GPU-side torch NCCL "
         "barrier (EP_KINETO_BARRIER=nccl), which needs no CUPTI build.",
     )
-    # NCCL-EP JIT knobs (defaults match the in-tree build; used by nccl/all).
+    # NCCL-EP JIT knobs: nccl.ep runtime-compiles (JITs) its device kernels on first use and
+    # must be told where the device sources (device/*.cuh) and NCCL public headers live. The
+    # defaults match the standard in-tree install, so these are normally unused; they are exposed
+    # only so a non-standard NCCL-EP install location can be overridden without editing the script.
     p.add_argument(
         "--nccl-jit-source-dir",
         default=os.environ.get("NCCL_EP_JIT_SOURCE_DIR", "/opt/microsoft/mrc/ep/nccl/contrib/nccl_ep"),
@@ -209,423 +225,6 @@ def parse_args() -> argparse.Namespace:
     if args.dispatch_dtype == "fp8_e4m3" and args.backend in ("nccl", "all"):
         raise SystemExit("--dispatch-dtype fp8_e4m3 is only supported by the mscclpp backend; use --backend mscclpp")
     return args
-
-
-# ----------------------------------------------------------------------------
-# MPI bootstrap shared by both backends.
-# ----------------------------------------------------------------------------
-def init_mpi():
-    comm = MPI.COMM_WORLD
-    rank = comm.Get_rank()
-    size = comm.Get_size()
-    local_rank = int(os.environ.get("OMPI_COMM_WORLD_LOCAL_RANK", rank))
-    torch.cuda.set_device(local_rank)
-    return comm, rank, size, local_rank
-
-
-def _ensure_torch_dist(comm, rank, num_ranks):
-    """Lazily initialize the default torch.distributed NCCL group alongside MPI
-    (idempotent). MPI supplies the rendezvous (rank-0 IP + port broadcast). Both
-    the kineto GPU barrier and the DeepEP backend reuse this single group.
-    Returns the world ProcessGroup."""
-    import torch.distributed as dist
-
-    if not dist.is_initialized():
-        addr = None
-        if rank == 0:
-            addr = os.environ.get("MASTER_ADDR")
-            if not addr:
-                import socket
-
-                _s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-                try:
-                    _s.connect(("10.255.255.255", 1))  # no packets sent; picks outbound iface IP
-                    addr = _s.getsockname()[0]
-                finally:
-                    _s.close()
-        addr = comm.bcast(addr, root=0)
-        port = comm.bcast(int(os.environ.get("MASTER_PORT", "29700")), root=0)
-        os.environ["MASTER_ADDR"] = addr
-        os.environ["MASTER_PORT"] = str(port)
-        os.environ["WORLD_SIZE"] = str(num_ranks)
-        os.environ["RANK"] = str(rank)
-        import datetime as _dt
-
-        dist.init_process_group(
-            backend="nccl", init_method="env://", world_size=num_ranks, rank=rank, timeout=_dt.timedelta(seconds=120)
-        )
-    return dist.distributed_c10d._get_default_group()
-
-
-def _init_torch_nccl(comm, rank, num_ranks, local_rank):
-    """Return a zero-arg GPU-side barrier (torch NCCL all_reduce) for the kineto
-    timing loop -- aligns ranks on-device, much tighter than an MPI host barrier."""
-    import torch.distributed as dist
-
-    _ensure_torch_dist(comm, rank, num_ranks)
-    _sync = torch.ones(1, dtype=torch.float, device="cuda")
-
-    def _barrier():
-        dist.all_reduce(_sync)
-
-    _barrier()
-    torch.cuda.synchronize()
-    return _barrier
-
-
-def _mpi_stats(comm, avg: float, mn: float, mx: float, num_ranks: int):
-    """Cross-rank reduction mirroring printLowLatencyResults: avg=mean of per-rank
-    avgs, min=global MIN, max=global MAX."""
-    g_avg = comm.allreduce(avg, op=MPI.SUM) / num_ranks
-    g_min = comm.allreduce(mn, op=MPI.MIN)
-    g_max = comm.allreduce(mx, op=MPI.MAX)
-    return g_avg, g_min, g_max
-
-
-# ----------------------------------------------------------------------------
-# Routing inputs — shared by both backends so the comparison is apples-to-apples.
-# BF16 tokens + top-k routing setup.
-# ----------------------------------------------------------------------------
-def make_inputs(num_tokens, hidden, num_topk, num_experts, rank, seed):
-    torch.manual_seed(seed + rank)
-    rank_offset = 128
-    x = torch.ones((num_tokens, hidden), dtype=torch.bfloat16, device="cuda") * (rank - rank_offset)
-    x[:, -128:] = torch.arange(num_tokens, device="cuda").to(torch.bfloat16).view(-1, 1)
-    scores = torch.randn((num_tokens, num_experts), dtype=torch.float32, device="cuda").abs() + 1
-    topk_idx = torch.topk(scores, num_topk, dim=-1, largest=True, sorted=True)[1].to(torch.int64)
-    topk_weights = torch.randn((num_tokens, num_topk), dtype=torch.float32, device="cuda").abs()
-    # ep_bench byte accounting: num_valid_selections = count(topk_idx >= 0); every
-    # selection is valid here (a full LL load), so this equals num_tokens * top_k.
-    num_valid_selections = int((topk_idx >= 0).sum().item())
-    return x, topk_idx, topk_weights, num_valid_selections
-
-
-# ----------------------------------------------------------------------------
-# LL dtype / combine helpers (ported from test_low_latency_multirank.py).
-# ----------------------------------------------------------------------------
-def fp8_e4m3_block128_scales(x):
-    blocks = x.float().reshape(*x.shape[:-1], x.size(-1) // 128, 128)
-    max_abs = blocks.abs().amax(dim=-1).clamp_min(1e-4)
-    return max_abs / 448.0
-
-
-def simulated_gemm_output(dispatch_out):
-    """Simulate the downstream expert GEMM so combine consumes BF16 expert output:
-    identity for BF16 dispatch; dequantize (tokens * block_scales) for FP8_E4M3."""
-    if dispatch_out.quant is None:
-        return dispatch_out.tokens
-    tokens = dispatch_out.tokens
-    token_blocks = tokens.float().reshape(*tokens.shape[:-1], tokens.size(-1) // 128, 128)
-    return (token_blocks * dispatch_out.quant.block_scales.unsqueeze(-1)).reshape(tokens.shape).to(torch.bfloat16)
-
-
-def validate_combine_output_mpi(actual, expected, comm, *, exact):
-    """MPI analog of the test's validate_combine_output: global max abs diff plus a
-    cross-rank finiteness (and, for direct_send, bit-exactness) assertion."""
-    local_diff = float((actual.float() - expected.float()).abs().max().item())
-    global_diff = comm.allreduce(local_diff, op=MPI.MAX)
-    local_finite = int(torch.isfinite(actual).all().item())
-    assert comm.allreduce(local_finite, op=MPI.MIN) == 1, "LL combine output contains NaN or Inf"
-    if exact:
-        local_equal = int(torch.equal(actual, expected))
-        assert comm.allreduce(local_equal, op=MPI.MIN) == 1, f"LL direct-send combine not bit-exact; diff={global_diff}"
-    else:
-        assert global_diff <= 8.0, f"LL rank-local combine mismatch; max diff={global_diff}"
-    return global_diff
-
-
-# ============================================================================
-# Backend: mscclpp EP (MoECommunicator).
-# ============================================================================
-def setup_mscclpp(args, comm, rank, num_ranks, inputs):
-    from mscclpp import CommGroup
-    import mscclpp.ep as ep
-
-    x, topk_idx, topk_weights, _ = inputs
-    num_tokens, hidden = args.num_tokens, args.hidden
-    num_experts, num_topk = args.num_experts, args.num_topk
-    num_local_experts = num_experts // num_ranks
-
-    num_rdma_bytes = 0  # not exposed by current mscclpp API; 0 over the CUDA-IPC path
-    if rank == 0:
-        print(
-            f"[cfg] backend=mscclpp algorithm=LOW_LATENCY num_ranks={num_ranks} tokens/rank={num_tokens} "
-            f"hidden={hidden} num_experts={num_experts} top_k={num_topk} "
-            f"warmup={args.num_warmup} iters={args.num_iters} num_rdma_bytes={num_rdma_bytes}",
-            flush=True,
-        )
-
-    ep_group = CommGroup(mpi_comm=comm)
-    combine_mode = {
-        "rank_local_reduce": ep.CombineMode.RANK_LOCAL_REDUCE,
-        "direct_send": ep.CombineMode.DIRECT_SEND,
-    }[args.combine_mode]
-    dispatch_quant = ep.QuantConfig(format=ep.DispatchDataType.FP8_E4M3) if args.dispatch_dtype == "fp8_e4m3" else None
-    dispatch_dtype = torch.float8_e4m3fn if dispatch_quant is not None else torch.bfloat16
-    moe_comm = ep.MoECommunicator(
-        comm=ep_group,
-        num_experts=num_experts,
-        num_local_experts=num_local_experts,
-        hidden_size=hidden,
-        topk=num_topk,
-        max_tokens_per_rank=num_tokens,
-        mode=ep.MoEMode.LOW_LATENCY,
-        low_latency_combine_mode=combine_mode,
-        quant=dispatch_quant,
-    )
-    assert moe_comm.is_available()
-    if rank == 0:
-        print(
-            f"[cfg] mscclpp MoECommunicator is_internode={moe_comm.is_internode()} "
-            f"dispatch_dtype={args.dispatch_dtype} combine_mode={args.combine_mode} cuda_graph={args.cuda_graph}",
-            flush=True,
-        )
-        if args.ep_layout == "rank_major":
-            print(
-                "[cfg] mscclpp ep_layout=rank_major requested but unsupported: LOW_LATENCY mode is "
-                "expert-major only ([num_local_experts, ...]); keeping expert_major "
-                "(rank/token-major is a HIGH_THROUGHPUT-mode feature)",
-                flush=True,
-            )
-
-    # Hoist output tensors out of the timed loop (the communicator owns its
-    # src_info/layout_range/count buffers internally).
-    output_buffer = torch.empty(
-        (num_local_experts, num_ranks * num_tokens, hidden), dtype=dispatch_dtype, device="cuda"
-    )
-    out = torch.empty((num_tokens, hidden), dtype=torch.bfloat16, device="cuda")
-
-    def _dispatch():
-        # Full (send+recv) LL dispatch inline on the stream; returns (dispatch_out, handle).
-        return moe_comm.dispatch(x, topk_idx, topk_weights, output_buffer=output_buffer)
-
-    def _combine(dispatch_out, handle):
-        # Feed BF16 expert output (identity for BF16, dequantized for FP8) into combine.
-        moe_comm.combine(simulated_gemm_output(dispatch_out), handle, out=out)
-
-    # Optional one-time correctness check (mirrors test_low_latency_multirank).
-    if args.validate:
-        v_dispatch_out, v_handle = _dispatch()
-        v_out = torch.empty_like(out)
-        moe_comm.combine(simulated_gemm_output(v_dispatch_out), v_handle, out=v_out)
-        torch.cuda.synchronize()
-        if dispatch_quant is None:
-            expected_f = torch.zeros_like(x, dtype=torch.float32)
-            x_f = x.float()
-            for j in range(num_topk):
-                weight_j = topk_weights[:, j].masked_fill(topk_idx[:, j] < 0, 0.0).view(-1, 1)
-                expected_f = torch.addcmul(expected_f, x_f, weight_j)
-            gdiff = validate_combine_output_mpi(
-                v_out, expected_f.to(torch.bfloat16), comm, exact=args.combine_mode == "direct_send"
-            )
-            if rank == 0:
-                print(f"[validate] mscclpp combine OK max|got-expected|={gdiff:.4e}", flush=True)
-        else:
-            assert torch.isfinite(v_out).all().item(), "FP8 LL combine produced NaN/Inf"
-            if rank == 0:
-                print("[validate] mscclpp FP8 combine finite OK", flush=True)
-
-    state = {"moe": moe_comm, "obuf": output_buffer, "out": out, "grp": ep_group}
-
-    if args.cuda_graph:
-        # Prime once, then capture dispatch and combine as separate CUDA graphs so
-        # the shared timed loop keeps its per-phase dispatch/combine events.
-        prime_out, prime_handle = _dispatch()
-        _combine(prime_out, prime_handle)
-        torch.cuda.synchronize()
-
-        g_dispatch = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g_dispatch):
-            g_dispatch_out, g_handle = moe_comm.dispatch(x, topk_idx, topk_weights, output_buffer=output_buffer)
-        g_combine = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g_combine):
-            moe_comm.combine(simulated_gemm_output(g_dispatch_out), g_handle, out=out)
-        state["graphs"] = (g_dispatch, g_combine)
-
-        def dispatch_fn():
-            g_dispatch.replay()
-            return (g_dispatch_out, g_handle)
-
-        def combine_fn(dout):
-            g_combine.replay()
-
-    else:
-
-        def dispatch_fn():
-            return _dispatch()
-
-        def combine_fn(dout):
-            dispatch_out, handle = dout
-            _combine(dispatch_out, handle)
-
-    def teardown():
-        state.clear()
-        gc.collect()
-        torch.cuda.synchronize()
-
-    return dispatch_fn, combine_fn, teardown
-
-
-# ============================================================================
-# Backend: NVIDIA NCCL-EP (nccl.ep Group/Handle).
-# ============================================================================
-def setup_nccl(args, comm, rank, num_ranks, inputs):
-    os.environ.setdefault("NCCL_EP_JIT_SOURCE_DIR", args.nccl_jit_source_dir)
-    os.environ.setdefault("NCCL_EP_JIT_BUILD_INCLUDE_DIR", args.nccl_jit_include_dir)
-
-    import nccl.core as nccl_core
-    import nccl.ep as nccl_ep
-
-    x, topk_idx, topk_weights, _ = inputs
-    num_tokens, hidden = args.num_tokens, args.hidden
-    num_experts, num_topk = args.num_experts, args.num_topk
-    num_local_experts = num_experts // num_ranks
-
-    if rank == 0:
-        print(
-            f"[cfg] backend=nccl algorithm=LOW_LATENCY num_ranks={num_ranks} tokens/rank={num_tokens} "
-            f"hidden={hidden} num_experts={num_experts} top_k={num_topk} "
-            f"warmup={args.num_warmup} iters={args.num_iters}",
-            flush=True,
-        )
-
-    # NCCL communicator: rank 0 makes a unique id, broadcast over MPI.
-    uid = nccl_core.get_unique_id() if rank == 0 else None
-    uid = comm.bcast(uid, root=0)
-    ncomm = nccl_core.Communicator.init(nranks=num_ranks, rank=rank, unique_id=uid)
-
-    config = nccl_ep.GroupConfig(
-        algorithm=nccl_ep.Algorithm.LOW_LATENCY,
-        num_experts=num_experts,
-        max_dispatch_tokens_per_rank=num_tokens,
-        max_recv_tokens_per_rank=num_tokens * num_ranks,
-        max_token_bytes=hidden * 2,  # BF16
-        alloc=nccl_ep.AllocConfig(),  # default cudaMalloc/cudaFree
-    )
-    ep_group = nccl_ep.Group.create(ncomm, config)
-
-    stream_ptr = torch.cuda.current_stream().cuda_stream
-
-    # Received-token layout: native/expert_major -> EXPERT_MAJOR (the historical default),
-    # rank_major -> RANK_MAJOR. NCCL-EP supports both natively via the handle Layout.
-    nccl_layout = nccl_ep.Layout.RANK_MAJOR if args.ep_layout == "rank_major" else nccl_ep.Layout.EXPERT_MAJOR
-    if rank == 0:
-        _lname = getattr(nccl_layout, "name", str(nccl_layout))
-        print(f"[cfg] nccl ep_layout={args.ep_layout} -> Layout.{_lname}", flush=True)
-
-    # Routing is encoded in the handle at create time (topk_idx is fixed for the run).
-    topk_idx_t = nccl_ep.Tensor(topk_idx)
-    ep_handle = ep_group.create_handle(
-        nccl_layout,
-        topk_idx_t,
-        layout_info=None,
-        config=nccl_ep.HandleConfig(),
-        stream=stream_ptr,
-    )
-
-    # Pre-allocated EP tensors (hoisted out of the timed loop). The recv-buffer shape
-    # and layout counters differ by layout: EXPERT_MAJOR groups received tokens per
-    # local expert ([num_local_experts, num_ranks*num_tokens, hidden] + per-expert
-    # counters); RANK_MAJOR groups them per source rank ([num_ranks, num_tokens, hidden]
-    # + per-source-rank counters, which the LL kernel asserts has leading dim == nRanks).
-    if nccl_layout == nccl_ep.Layout.RANK_MAJOR:
-        recv = torch.empty((num_ranks, num_tokens, hidden), dtype=torch.bfloat16, device="cuda")
-        recv_count = torch.empty((num_ranks,), dtype=torch.int32, device="cuda")
-    else:
-        recv = torch.empty((num_local_experts, num_ranks * num_tokens, hidden), dtype=torch.bfloat16, device="cuda")
-        recv_count = torch.empty((num_local_experts,), dtype=torch.int32, device="cuda")
-    out = torch.empty((num_tokens, hidden), dtype=torch.bfloat16, device="cuda")
-
-    x_t = nccl_ep.Tensor(x)
-    recv_t = nccl_ep.Tensor(recv)
-    recv_count_t = nccl_ep.Tensor(recv_count)
-    out_t = nccl_ep.Tensor(out)
-    topk_weights_t = nccl_ep.Tensor(topk_weights)
-
-    if nccl_layout == nccl_ep.Layout.RANK_MAJOR:
-        # RANK_MAJOR dispatch carries the per-token weights through as well (the LL
-        # kernel asserts topk_weights_in != NULL), and writes them into a matching
-        # per-source-rank recv buffer.
-        recv_weights = torch.empty((num_ranks, num_tokens, num_topk), dtype=torch.float32, device="cuda")
-        recv_weights_t = nccl_ep.Tensor(recv_weights)
-        recv_idx = torch.empty((num_ranks, num_tokens, num_topk), dtype=torch.int32, device="cuda")
-        recv_idx_t = nccl_ep.Tensor(recv_idx)
-        dispatch_inputs = nccl_ep.DispatchInputs(tokens=x_t, topk_weights=topk_weights_t)
-        dispatch_outputs = nccl_ep.DispatchOutputs(tokens=recv_t, topk_weights=recv_weights_t, topk_idx=recv_idx_t)
-        dispatch_layout = nccl_ep.LayoutInfo(src_rank_counters=recv_count_t)
-    else:
-        dispatch_inputs = nccl_ep.DispatchInputs(tokens=x_t)
-        dispatch_outputs = nccl_ep.DispatchOutputs(tokens=recv_t)
-        dispatch_layout = nccl_ep.LayoutInfo(expert_counters=recv_count_t)
-    dispatch_config = nccl_ep.DispatchConfig()
-    combine_inputs = nccl_ep.CombineInputs(tokens=recv_t)
-    combine_outputs = nccl_ep.CombineOutputs(tokens=out_t, topk_weights=topk_weights_t)
-    combine_config = nccl_ep.CombineConfig()
-
-    def _dispatch(stream):
-        # Pure on-stream kernel launch; routing is baked into ep_handle, all
-        # tensors are pre-allocated, so nothing runs host-side per call.
-        ep_handle.dispatch(
-            dispatch_inputs,
-            dispatch_outputs,
-            layout_info=dispatch_layout,
-            config=dispatch_config,
-            stream=stream,
-        )
-
-    def _combine(stream):
-        ep_handle.combine(combine_inputs, combine_outputs, config=combine_config, stream=stream)
-
-    graphs = []
-
-    if args.cuda_graph:
-        if rank == 0:
-            print("[cfg] nccl cuda_graph=True (capturing dispatch/combine)", flush=True)
-        # NCCL-EP dispatch/combine are a PAIRED collective (combine consumes the
-        # peer data produced by the matching dispatch); main() forces the paired
-        # kineto pass (EP_KINETO_SEPARATE=0) so every captured combine follows a
-        # captured dispatch instead of spinning on a stale peer receive.
-        # Prime once so any lazy JIT / autotune settles before capture.
-        _dispatch(stream_ptr)
-        _combine(stream_ptr)
-        torch.cuda.synchronize()
-
-        # Capture on the graph's capture stream: NCCL-EP takes an explicit stream
-        # pointer, so it must be re-fetched inside the capture context (the default
-        # stream_ptr would launch off the capture stream and break capture).
-        g_dispatch = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g_dispatch):
-            _dispatch(torch.cuda.current_stream().cuda_stream)
-        g_combine = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g_combine):
-            _combine(torch.cuda.current_stream().cuda_stream)
-        graphs = [g_dispatch, g_combine]
-
-        def dispatch_fn():
-            g_dispatch.replay()
-            return None
-
-        def combine_fn(_dout):
-            g_combine.replay()
-
-    else:
-
-        def dispatch_fn():
-            _dispatch(stream_ptr)
-            return None
-
-        def combine_fn(_dout):
-            _combine(stream_ptr)
-
-    def teardown():
-        graphs.clear()  # drop graph refs before the handle they captured is destroyed
-        ep_handle.destroy()
-        ep_group.destroy()
-        ncomm.destroy()
-        gc.collect()
-        torch.cuda.synchronize()
-
-    return dispatch_fn, combine_fn, teardown
 
 
 # ============================================================================
@@ -1004,315 +603,6 @@ class _InProcCupti:
 
     def count(self, substr):
         return int(self.lib.kt_get_count(substr.encode()))
-
-
-# ============================================================================
-# Backend: DeepEP V2 (deepseek-ai/DeepEP, ElasticBuffer low-latency).
-# ============================================================================
-def setup_deepep(args, comm, rank, num_ranks, inputs):
-    """DeepEP V2 low-latency dispatch/combine via `deep_ep.ElasticBuffer`, wired
-    the same way as the mscclpp / NCCL-EP backends: return (dispatch_fn,
-    combine_fn, teardown). DeepEP needs a torch.distributed NCCL group (reused
-    from `_ensure_torch_dist`) and, for a same-rack NVLink run, `EP_DISABLE_GIN=1`.
-    The dispatch handle (routing) is fixed for the run, so we dispatch once to
-    obtain the handle + combine input and then replay dispatch/combine in the
-    timed loop -- dispatch_impl(+copy) and combine_impl(+reduce) are what the
-    kineto timer buckets by the "dispatch"/"combine" substrings."""
-    import deep_ep
-
-    os.environ.setdefault("EP_DISABLE_GIN", "1")
-    group = _ensure_torch_dist(comm, rank, num_ranks)
-
-    x, topk_idx, topk_weights, _ = inputs
-    num_tokens, hidden = args.num_tokens, args.hidden
-    num_experts, num_topk = args.num_experts, args.num_topk
-
-    if rank == 0:
-        print(
-            f"[cfg] backend=deepep ElasticBuffer(LOW_LATENCY) num_ranks={num_ranks} tokens/rank={num_tokens} "
-            f"hidden={hidden} num_experts={num_experts} top_k={num_topk} "
-            f"warmup={args.num_warmup} iters={args.num_iters}",
-            flush=True,
-        )
-
-    buffer = deep_ep.ElasticBuffer(
-        group,
-        num_max_tokens_per_rank=num_tokens,
-        hidden=hidden,
-        allow_hybrid_mode=1,
-        allow_multiple_reduction=1,
-        explicitly_destroy=True,
-    )
-
-    # Received-token layout: rank_major (plain, do_expand=False -- the current default)
-    # or expert_major (do_expand=True -- the expanding layout, one slot per expert per
-    # token). 'native' keeps the plain rank-major layout.
-    do_expand = args.ep_layout == "expert_major"
-    if rank == 0:
-        _lay = "expert_major (do_expand)" if do_expand else "rank_major (plain)"
-        print(f"[cfg] deepep ep_layout={args.ep_layout} -> {_lay}", flush=True)
-
-    # DeepEP dispatch args (BF16; non-cached so dispatch_impl + copy epilogue run).
-    dispatch_args = dict(
-        x=x,
-        topk_idx=topk_idx,
-        topk_weights=topk_weights,
-        num_experts=num_experts,
-        num_max_tokens_per_rank=num_tokens,
-        do_cpu_sync=True,
-        do_expand=do_expand,
-        do_zero_padding=do_expand,
-    )
-
-    # Prime once to obtain the handle (routing) and the received-token count so we
-    # can size the combine input. Build a realistic BF16 combine input in the
-    # received layout (the role of `simulated_gemm_output`): random values only in
-    # the valid received-token slots, matching test_ep.py's `input_for_combine`.
-    #
-    # NOTE: the PLAIN (rank-major) path uses the per-expert combine; the expanded
-    # (expert-major) path runs DeepEP's reduce/expand combine, which processes more
-    # rows in the reduce epilogue (measured more expensive on this uniform workload:
-    # 1n 74 vs 46, 4n 188 vs 136, 8n 199 vs 120 us). Both are exercised via --ep-layout.
-    recv_x, recv_topk_idx, recv_topk_weights, handle, _ = buffer.dispatch(**dispatch_args)
-    recv_x_bf16 = recv_x[0] if isinstance(recv_x, (tuple, list)) else recv_x
-    input_for_combine = torch.empty_like(recv_x_bf16, dtype=torch.bfloat16)
-    input_for_combine.normal_(0.0, 0.1)
-    if not do_expand:
-        # Plain layout: zero the invalid received-token slots (rank-major padding).
-        num_recv_tokens = int(handle.psum_num_recv_tokens_per_scaleup_rank[-1].item())
-        if num_recv_tokens < input_for_combine.shape[0]:
-            input_for_combine[num_recv_tokens:] = 0
-
-    combine_args = dict(
-        x=input_for_combine,
-        topk_weights=recv_topk_weights,
-        handle=handle,
-    )
-
-    # DeepEP native GPU barrier (comm-stream, sequential) -- aligns its dispatch/
-    # combine recv-spin far more tightly than a generic all_reduce; this is what
-    # DeepEP's own test_ep.py uses for kineto profiling.
-    def deepep_barrier():
-        buffer.barrier(use_comm_stream=True, with_cpu_sync=False, sequential=True)
-
-    graphs = []
-
-    # DeepEP CUDA-graph capture only works on the intranode NVLink path. On the
-    # internode scale-out path DeepEP's symmetric-memory kernels crash under graph
-    # capture (CUDA 719 in symmetric.hpp), so restrict capture to a single node.
-    local_world = int(os.environ.get("MSCCLPP_EP_LOCAL_WORLD_SIZE", "0") or "0")
-    deepep_can_graph = args.cuda_graph and (local_world <= 0 or num_ranks <= local_world)
-
-    if args.cuda_graph and not deepep_can_graph and rank == 0:
-        print(
-            "[cfg] deepep cuda_graph requested but disabled: internode scale-out is "
-            "not graph-capturable (symmetric-memory); using eager dispatch/combine",
-            flush=True,
-        )
-
-    if deepep_can_graph:
-        if rank == 0:
-            print("[cfg] deepep cuda_graph=True (cached dispatch, do_cpu_sync=False)", flush=True)
-        # The non-cached dispatch above did a CPU sync to size the layout; that is
-        # illegal inside a CUDA graph. Replay the CACHED dispatch instead: pass the
-        # primed handle (topk_idx reused from it), which forces do_cpu_sync=False and
-        # skips the host-side count read, leaving a pure on-stream kernel launch.
-        cached_dispatch_args = dict(
-            x=x,
-            handle=handle,
-            num_experts=num_experts,
-            num_max_tokens_per_rank=num_tokens,
-        )
-        # Prime the cached path once so any lazy alloc/autotune settles before capture.
-        buffer.dispatch(**cached_dispatch_args)
-        buffer.combine(**combine_args)
-        torch.cuda.synchronize()
-
-        g_dispatch = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g_dispatch):
-            buffer.dispatch(**cached_dispatch_args)
-        g_combine = torch.cuda.CUDAGraph()
-        with torch.cuda.graph(g_combine):
-            buffer.combine(**combine_args)
-        graphs = [g_dispatch, g_combine]
-
-        def dispatch_fn():
-            g_dispatch.replay()
-            return None
-
-        def combine_fn(_dout):
-            g_combine.replay()
-
-    else:
-
-        def dispatch_fn():
-            buffer.dispatch(**dispatch_args)
-            return None
-
-        def combine_fn(_dout):
-            buffer.combine(**combine_args)
-
-    def teardown():
-        graphs.clear()  # drop graph refs before the buffer they captured is destroyed
-        try:
-            buffer.destroy()
-        except Exception:
-            pass
-        gc.collect()
-        torch.cuda.synchronize()
-
-    return dispatch_fn, combine_fn, teardown, deepep_barrier
-
-
-# ============================================================================
-# Backend: FlashInfer (flashinfer.comm.trtllm_moe_alltoall, MoeAlltoAll / MNNVL).
-# ============================================================================
-def setup_flashinfer(args, comm, rank, num_ranks, inputs):
-    """FlashInfer throughput MoE all-to-all via `flashinfer.comm.trtllm_moe_alltoall.
-    MoeAlltoAll` over the MNNVL fabric, wired like the other backends: return
-    (dispatch_fn, combine_fn, teardown). No native barrier, so the harness aligns
-    ranks with its GPU-side torch NCCL all_reduce (nccl_barrier). The kernels are
-    named ``moeA2ADispatchKernel`` / ``moeA2ACombineKernel`` (+ prepare kernels),
-    which the kineto timer buckets by the "dispatch"/"combine" substrings.
-
-    Env EP_FLASHINFER_GPUS_PER_NODE (default 4) sets the MNNVL Mapping layout."""
-    from flashinfer.comm.mapping import Mapping
-    import flashinfer.comm.trtllm_moe_alltoall as a2a
-
-    x, topk_idx, _topk_weights, _ = inputs
-    num_tokens, hidden = args.num_tokens, args.hidden
-    num_experts, num_topk = args.num_experts, args.num_topk
-    gpus_per_node = int(os.environ.get("EP_FLASHINFER_GPUS_PER_NODE", "4"))
-    dev = x.device
-    dtype = torch.bfloat16
-
-    if rank == 0:
-        print(
-            f"[cfg] backend=flashinfer MoeAlltoAll(MNNVL) num_ranks={num_ranks} tokens/rank={num_tokens} "
-            f"hidden={hidden} num_experts={num_experts} top_k={num_topk} "
-            f"warmup={args.num_warmup} iters={args.num_iters}",
-            flush=True,
-        )
-        if args.ep_layout == "expert_major":
-            print(
-                "[cfg] flashinfer ep_layout=expert_major requested but unsupported: "
-                "MoeAlltoAll dispatch is fixed rank-major [ep_size, tokens, hidden]; "
-                "keeping rank_major (expert grouping is a downstream model op, not part of the A2A)",
-                flush=True,
-            )
-
-    # EP is expressed as tp_size = world with moe_ep_size = world, moe_tp_size = 1
-    # (Mapping requires world_size == tp*pp*cp).
-    mapping = Mapping(
-        world_size=num_ranks,
-        rank=rank,
-        gpus_per_node=gpus_per_node,
-        tp_size=num_ranks,
-        moe_tp_size=1,
-        moe_ep_size=num_ranks,
-    )
-    moe = a2a.MoeAlltoAll(
-        mapping,
-        max_num_tokens=num_tokens,
-        top_k=num_topk,
-        num_experts=num_experts,
-        hidden_size=hidden,
-    )
-
-    # Shared routing inputs: FlashInfer wants int32 expert ids [num_tokens, top_k]
-    # and the hidden-state payload in BF16. Reuse the harness's topk_idx / x so the
-    # workload matches the other backends.
-    token_selected_experts = topk_idx[:, :num_topk].to(torch.int32).contiguous()
-    hidden_payload = x.to(dtype).contiguous()
-    # Combine payload lives in the received layout [ep_size, max_tokens, hidden].
-    combine_payload = torch.randn(num_ranks, num_tokens, hidden, generator=None, device=dev, dtype=dtype)
-
-    # FlashInfer's MoeAlltoAll is STATEFUL: dispatch() sets phase="dispatched" and
-    # combine() requires it then resets to "idle". The harness's paired loop calls
-    # dispatch_fn then combine_fn in order (phase satisfied). Each op begins with an
-    # MPI barrier so all ranks enter the kernel roughly aligned -- FlashInfer's
-    # in-kernel peer-readiness spin otherwise deadlocks multi-node when ranks drift
-    # (this is exactly how the standalone FlashInfer bench aligns ranks). The
-    # host barrier is safe here because FlashInfer dispatch/combine are independent
-    # (unlike DeepEP, whose paired ops share symmetric-memory state).
-    # EP_KINETO_SEPARATE is forced off for FlashInfer so the timer replays the
-    # dispatch->combine pair in order (a lone combine would trip the phase assert).
-    os.environ["EP_KINETO_SEPARATE"] = "0"
-
-    graphs = []
-
-    def _dispatch():
-        moe.dispatch(token_selected_experts, [hidden_payload], num_tokens)
-
-    def _combine():
-        moe.combine(combine_payload, num_tokens)
-
-    captured = False
-    if args.cuda_graph:
-        # FlashInfer's dispatch/combine each begin with an MPI host barrier to align
-        # ranks -- but an MPI barrier cannot live inside a CUDA graph. So capture ONLY
-        # the moe kernels and keep the barrier OUTSIDE, replaying: barrier -> replay.
-        # Capture is best-effort: FlashInfer's stateful phase / in-kernel peer spin may
-        # not be graph-capturable on every build, so on failure we rebuild the (now
-        # possibly mid-phase) communicator and fall back to the direct barrier+launch.
-        try:
-            # Prime one full pair so phase returns to idle and lazy work settles.
-            _dispatch()
-            _combine()
-            torch.cuda.synchronize()
-            g_dispatch = torch.cuda.CUDAGraph()
-            with torch.cuda.graph(g_dispatch):
-                _dispatch()
-            g_combine = torch.cuda.CUDAGraph()
-            with torch.cuda.graph(g_combine):
-                _combine()
-            torch.cuda.synchronize()
-            graphs = [g_dispatch, g_combine]
-            captured = True
-            if rank == 0:
-                print("[cfg] flashinfer cuda_graph=True (kernels captured; MPI barrier kept outside)", flush=True)
-        except Exception as e:  # noqa: BLE001 - capturability is an external-library boundary
-            if rank == 0:
-                print(f"[cfg] flashinfer cuda_graph capture failed ({e}); falling back to eager", flush=True)
-            # Rebuild to clear any partially-advanced phase state.
-            moe = a2a.MoeAlltoAll(
-                mapping,
-                max_num_tokens=num_tokens,
-                top_k=num_topk,
-                num_experts=num_experts,
-                hidden_size=hidden,
-            )
-        comm.Barrier()
-
-    if captured:
-        g_dispatch, g_combine = graphs
-
-        def dispatch_fn():
-            comm.Barrier()
-            g_dispatch.replay()
-            return None
-
-        def combine_fn(_dout):
-            comm.Barrier()
-            g_combine.replay()
-
-    else:
-
-        def dispatch_fn():
-            comm.Barrier()
-            moe.dispatch(token_selected_experts, [hidden_payload], num_tokens)
-            return None
-
-        def combine_fn(_dout):
-            comm.Barrier()
-            moe.combine(combine_payload, num_tokens)
-
-    def teardown():
-        graphs.clear()
-        gc.collect()
-        torch.cuda.synchronize()
-
-    return dispatch_fn, combine_fn, teardown
 
 
 _SETUP = {"mscclpp": setup_mscclpp, "nccl": setup_nccl, "deepep": setup_deepep, "flashinfer": setup_flashinfer}
